requirements.txt
streamlit>=1.36.0
pandas>=2.0.0
PyPDF2>=3.0.0
spacy>=3.7.0
python-docx>=1.0.0